models/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pandas>=2.2.0
numpy>=2.2.6
pandas-ta>=0.4.71b0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
